refactor: put all packages in internal to server directory;
