+++ cmd/server/main.go
@@ -6,13 +6,13 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/ghazlabs/idn-remote-entry/internal/core"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/notifier"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/hqloc"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/parser"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/storage"
-	"github.com/ghazlabs/idn-remote-entry/internal/driver"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/core"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/notifier"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/hqloc"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/parser"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/storage"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driver"
 	"github.com/go-resty/resty/v2"
 	"github.com/openai/openai-go"
 	"github.com/openai/openai-go/option"
