+++ internal/server/core/model.go
