+++ internal/server/core/service.go
