+++ internal/server/driven/resolver/hqloc/locator.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/util"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/util"
 	"github.com/go-resty/resty/v2"
 	"github.com/openai/openai-go"
 	"gopkg.in/validator.v2"
