+++ internal/server/driven/resolver/hqloc/model.go
