+++ internal/server/driven/resolver/model.go
@@ -1,6 +1,6 @@
 package resolver
 
-import "github.com/ghazlabs/idn-remote-entry/internal/core"
+import "github.com/ghazlabs/idn-remote-entry/internal/server/core"
 
 type VacancyInfo struct {
 	JobTitle         string   `json:"job_title" jsonschema_description:"Title of the job taken from the vacancy description, if not found then it should be empty"`
