+++ internal/server/driven/resolver/parser/greenhouse.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 
 	"github.com/PuerkitoBio/goquery"
-	"github.com/ghazlabs/idn-remote-entry/internal/core"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/util"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/core"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/util"
 	"github.com/go-resty/resty/v2"
 	"github.com/openai/openai-go"
 	"gopkg.in/validator.v2"
