+++ internal/server/driven/resolver/vacancy_test.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/hqloc"
-	"github.com/ghazlabs/idn-remote-entry/internal/driven/resolver/parser"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/hqloc"
+	"github.com/ghazlabs/idn-remote-entry/internal/server/driven/resolver/parser"
 	"github.com/ghazlabs/idn-remote-entry/internal/testutil"
 	"github.com/go-resty/resty/v2"
 	"github.com/openai/openai-go"
